+++ source/libnormaliz/cone.h
@@ -494,7 +494,7 @@ class Cone {
 
     renf_elem_class getFieldElemConeProperty(ConeProperty::Enum property);
 
-    size_t getMachineIntegerConeProperty(ConeProperty::Enum property);
+    long getMachineIntegerConeProperty(ConeProperty::Enum property);
 
     bool getBooleanConeProperty(ConeProperty::Enum property);
